Expose View definition internals for easier debugging Turns out there is no good reason for treetop.View to be an interface
