+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 ## [0.2.0] - 2020-01-26
 
-
 ### Breaking API Changes
 
 Taking Treetop from POC to Alpha gives me an opportunity to
@@ -11,6 +10,7 @@ execute on a wishlist of API changes.
 - Rename `treetop.TreetopWriter` interface to `treetop.Writer` to conform to naming guidelines
 - Remove `treetop.Test`, testing recipes and resources belong elsewhere
 - Split `treetop.Writer` function into `treetop.NewPartialWriter` and `treetop.NewFragmentWriter` and remove the confusing `isPartial` flag
+- Change `treetop.View` from an interface to a struct and expose internals to make debugging easier
 
 
 #### Defining a page with views
